+++ R/layers.R
@@ -100,6 +100,8 @@ layer3D.callback <- function(fun, args = list()) {
         do.call(layer$fun, c(list(context), layer$args))
     } else if (layer$type == "axes") {
         .draw.axes.layer(layer, context)
+    } else if (layer$type == "mesh") {
+        .draw.mesh.layer(layer, context)
     } else if (layer$type == "edges") {
         .indices(as.vector(layer$edges), nrow(X), "edges")
         if (!is.null(layer$path)) .indices(layer$path, nrow(X), "path")